Like button with counter done

+++ app/controllers/likes_controller.rb
@@ -5,7 +5,7 @@ class LikesController < ApplicationController
   end
 
   def destroy
-    @post = current_user.posts.find_by(id: params[:post_id])
-    current_user.likes.where(post_id: @post.id).delete_all
+    @post = Post.find(params[:post_id])
+    current_user.likes.where(post_id: @post.id, user_id: current_user.id).delete_all
   end
 end
